requirements.txt
fastapi
uvicorn
pydantic
httpx
python-dotenv
python-jose
passlib
bcrypt
python-multipart
